% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsii.R
\name{get_nsii}
\alias{get_nsii}
\title{Download specimen data from the NSII}
\usage{
get_nsii(query, offset = 0)
}
\arguments{
\item{query}{\code{character} Only single Latin name is queried.}

\item{offset}{\code{numeric} Query page number.The default value is 0, which corresponds to the first page, each query returns 30.}
}
\value{
list
}
\description{
The result returns the detailed information of the queried species, including the scientific name, the Latin name, common name, description, etc.
}
\details{
Visit the website \url{http://www.nsii.org.cn} for more details.
}
\examples{
\dontrun{
result <- get_nsii(query = "Artemisia argyi",offset = 0)
str(result)
tibble::tibble(result$data)
}
}
\author{
Liuyong Ding
}
