% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occ.R
\name{get_occ}
\alias{get_occ}
\title{Download occurrence data from the NSII}
\usage{
get_occ(query)
}
\arguments{
\item{query}{\code{character} Only single Latin name is queried.}
}
\value{
list
}
\description{
The result returns the detailed information of the queried species, including the scientific name, the Latin name, common name, description, latitude and longitude, etc.
}
\details{
Visit the website \url{http://www.nsii.org.cn} for more details.
}
\examples{
\dontrun{
result <- get_occ(query = "Artemisia argyi")
str(result)
tibble::tibble(result$Specimens)
}
}
\author{
Liuyong Ding <ly_ding@126.com>
}
