\name{rnsii-package}
\alias{rnsii-package}
\alias{rnsii}
\docType{package}
\title{
Mining China biodiversity from the National Specimen Information Infrastructure
}
\description{
This package is designed for programatically download Chinese biodiversity data from the National Specimen Information Infrastructure <http://www.nsii.org.cn>.
}
\details{
\packageDESCRIPTION{rnsii}
\packageIndices{rnsii}
}

\author{Maintainer: \packageMaintainer{rnsii}}


\keyword{ package }
